fix: use generalized graph implementation
